Created 3 prompts
I created and tested the first prompt which was the user input prompt.
This was just a dialog box. Secondly I created a var and prompt for
searching the area of a rectangle. This was printed using console.log.
Lastly I commented out the console and created a dialog alert which
printed the area of a rectangle.