Assigned var and tested the script.
The first script was to assign and effectively test the console to
ensure the current age of sparky was being printed to the console. I
assigned var to both current human age, and dog years and added the dog
name. Next I tested the script by running it on the console. It was
effective. The current age was a product of current age (human years)
and dog years.