Slice Of Pie expressions worksheet
I completed the slice of pie portion of the worksheet by declaring
variables which would define the pies ordered, slices each pie had, and
the amount of guest. I used arithmetic expressions to figure out how
many slices each guest had and had a remainder. Everything is commented
out, tested in the console and then printed out onto an alert to the
screen.