Created a weekly budget average
I completed the average shopping bill assignment from the expressions
worksheet. Using an array, I created a var to add the array using the
index numbers. Then I checked it in the log to make sure it printed.
Next I created an alert which printed this information to the screen in
a dialog box. I also did the same for the average weekly bill using the
division operator.