Created the expressions assignment and have began coding
This calculator is for nurses to see if they are being overworked. I
have created the js and html file and tested it with an alert to make
sure it was running properly. I then created some user prompts to input
information. The next step was to output the days the user entered,
followed by a prompt question verifying the input. I am now working on
hour calculations. there is an array using numbers which will calculate
the amount of hours worked per day, and per month.