Week 3 Conditionals Worksheet first group completed
I completed the celsius to fahrenheit assignment. What I did was set
two variables. One for each temperature and degree units. Next I
created an if and else statement which would determine which message to
print to the console. Each step is commented with an explanation of
what I am doing. I also tested everything to make sure the code is
running properly.