Created and tested the user input portion of the assignment week 3
I created a code for the user input portion of week 3. I declared
variables for username and passwords as well as two prompts for
username and passwords. I then set one equal to and not equal to and
ran the script to make sure it was working properly. Everything is
commented appropriately.