Created concatenating strings, with 1 alert and two prompts
In order to test this calculator and make sure it is working, I created
two more prompts instructing the user to input the width and length of
their raised bed. The result was printed out to the console and an
alert was created on screen. I also added some more comments where they
were lacking to better explain the functionality of the calculator.