Finished the last part of the calculator and adjusted the seed count
I added ternary, and logical operators to the first part of the
calculator. This will determine if the raised garden bed will be ready
to plant crop in. I also added another variable which was implemented
into the relational operator statement and tested it to make sure it
was working. The last part contains to seed count which tells the user
what date or what month the crop can be expected.